Divide execution time with runtime in UI
We are counting suite execution time using a simple timer. This is fine,
however it's not the best way to do it when it comes to the tests
execution. So we need to split timing into 2 different parts: execution
time (effective kirk suite execution) and runtime (sum of all tests
executions).