Handle kernel tainted info in parallel testing
This patch provides a smarter way to recognize when kernel is tainted
during parallel testing. We keep track of the tainted status of the
kernel and return it if it has been fetched by other coroutines in the
meanwhile. In this way, we don't access to the same file multiple times
or even at the same time, creating dead-locks and program freeze.