Add overall test status in test results data
The overall test status is mapped inside the Results class with the name
of "status". This is a requirement that comes from LTP but it's shared
with other framework tests binaries as well. It's basically the overall
result of the test. We might have 10 tests inside a test binary, 9 PASS
but 1 SKIP and the overall result would be PASS. The flag handle these
types of cases.