Session restore support
Any time that a new session is executed, a file called 'executed' is
created inside the temporary folder. This file contains the list of
executed tests for a specific session in the format 'suite::test'.
When a session folder is passed via '--restore' argument, the session
will be restored starting from the next available test. This is useful
when kirk crashes for some reasons, for example when it's running
inside a VM that crashed due to a test execution.