Fix kirk causing RuntimeError on SIGINT
When SIGINT is received, kirk cancels all tasks but the current one.
This is fine, ideally, but practically when SIGINT is received event
loop has stopped already and we obtain RuntimeError. For this reason,
the best thing to do is to get all running tasks which asyncio is
providing and kill them without filtering asyncio.current_task