Fix qemu login prompt recognition
Most of the times we boot a new VM, login message shows '#' symbol
at the end. However, we are using '# ' (with space) to indicate
availability the user prompt. This can lead to an unrecognised
login prompt if message contain any color after the symbol.

With this patch, we take '#' symbol only into account, so we are able to
cover as much as cases as possible by default.