libkirk/events: register the event handler for suite_timeout
In our testing, we find when suite timeout happened, nothing is printed.
and it confuses us that many tests are skipped but no hint in the log.
We find the handler is there but not registered, so add it there.

With the line added, we'll have the log like below when suite timeout
happened:

Suite 'syscalls' timed out after 3600 seconds.

Signed-off-by: Chunyu Hu <chuhu@redhat.com>