options: add new --run-pattern
Add the --run-pattern|-S option to filter tests based on a specific
string pattern. This option accepts a list of strings, allowing users
to specify multiple patterns. A common usage example is as follows:

    kirk --run-pattern sendfile madvice

Each string is compared against the labels defined in the runtest file,
and only tests containing the specified patterns will be executed.

Signed-off-by: Andrea Cervesato <andrea.cervesato@suse.com>