ui: speed up user interface
Remove direct `print` usage, since it's a blocking I/O routine calling
the underlying `write()` syscall. Instead, run it via asyncio.to_thread.

The big challenge was given by the asynchronous nature of threads,
messing up the entire console once they complete in different times.
The way we resolved this is by introducing ordered coroutines
registration for single events, creating a `printf` event and using it
internally in the ConsoleUserInterface implementation.
In this way, first-in event will be printed as first-out event in the
print queue.

Signed-off-by: Andrea Cervesato <andrea.cervesato@suse.com>