ui: make parallel ui less interactive
User interface is quite difficult to track when we are running kirk
in a non-interactive mode (for example, via serial console). The idea is
to simplify the overall ui by printing at the beginning all tests which
will run in parallel. In this way we will reduce noise in the interface,
especially in the non-interactive mode which is our main target.

Fixes: https://github.com/linux-test-project/kirk/issues/58
Signed-off-by: Andrea Cervesato <andrea.cervesato@suse.com>