tests: xfail certain LTXSUT tests under github actions
When running under Github actions or any other containerized system,
SUT::stop() method might be unstable. This is the case of LTX, where
it seems that signals sent to the process, such as SIGKILL, are ignored
until the command will be completed. This is not happening in a local
environment and the test passes.

Signed-off-by: Andrea Cervesato <andrea.cervesato@suse.com>