ltp: don't override LTP_TIMEOUT_MUL if set by user
The Kirk framework was unconditionally calculating and setting the
LTP_TIMEOUT_MUL env variable based on the test_timeout parameter,
overriding any value that the user might have set beforehand.

This change ensures that if the LTP_TIMEOUT_MUL environment variable is
already set in the environment, the framework will not override it. This
allows users to explicitly control the timeout multiplier for LTP tests.

The automatic calculation based on test_timeout remains as a fallback
when LTP_TIMEOUT_MUL is not set by the user.

Signed-off-by: Li Wang <liwang@redhat.com>