framework: Framework is not longer a Plugin
The idea started with handling multiple frameworks under the kirk
umbrella, but it didn't go too well. The support requires too much
effort and we don't really want to stuck into legacy unsupported code or
weird generic frameworks in the Open Source world.

This patch will remove frameworks plugins once for all, so we can focus
on the one and only LTP framework.

The option `--framework` will also be removed and it will no longer
present.

Signed-off-by: Andrea Cervesato <andrea.cervesato@suse.com>